+++ Changelog.md
@@ -5,10 +5,27 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-- Added a Javadoc feature
-- Added a plugin for license management
-- Added JavaFX plugin
-- Added Gson dependency
+
+
+
+## [0.1.0] - 2022-02-25
+
+
+### Added
+ - Added JavaFX feature to make course editor application
+ - Added GSON to serialize and deserialize objects
+ - Added JavaDoc feature
+
+### Changed
+
+
+
+
+### Removed
+
+
+
+
 
 
 ## [0.0.1] - 2022-02-11